# Conflicts:
#	SCMM.Web.Client/SCMM.Web.Client.csproj

+++ SCMM.Shared.Data.Models/Extensions/StringExtensions.cs
@@ -112,7 +112,7 @@ namespace SCMM.Shared.Data.Models.Extensions
             return value.First().ToString().ToLower() + value.Substring(1);
         }
 
-        public static string Pluralise(this string value, int count = 0)
+        public static string Pluralise(this string value, long count = 0)
         {
             if (String.IsNullOrEmpty(value))
             {
